% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSequences.R
\name{readSequences}
\alias{readSequences}
\title{Read in sequence data}
\usage{
readSequences(filepath)
}
\arguments{
\item{filepath}{the filepath to the specific .txt file you are trying to read in}
}
\value{
a dataframe with two columns, column one is for the sequence name, column two is for the sequence data itself, each row is for each sequence entered
}
\description{
Read in sequence data
}
\examples{
\dontrun{
sequencesDF <- readSequences("path/to/sequences.txt")
}
}
