% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseSequence.R
\name{reverseSequence}
\alias{reverseSequence}
\title{Reverses a specified sequence}
\usage{
reverseSequence(df, name)
}
\arguments{
\item{df}{a dataframe with exactly 2 columns, typically generated by the readSequence function in this package}

\item{name}{the name of the sequence that needs to be reversed}
}
\value{
the "same" dataframe but with the specified dataframe reversed
}
\description{
Reverses a specified sequence
}
\examples{
sequences <- reverseSequence(df, "HisD TA1538 Mutant Genotype")

}
