% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateDNA.R
\name{translateDNA}
\alias{translateDNA}
\title{Translate a DNA sequence to a protein sequence}
\usage{
translateDNA(df, offset = 0)
}
\arguments{
\item{df}{a dataframe with exactly 2 columns, typically generated by the readSequence function in this package}

\item{offset}{a number 0,1,2 that determines the reading frame offset from the start of the}
}
\value{
a dataframe with 3 columns, name, DNA sequence, protein sequence
}
\description{
Translate a DNA sequence to a protein sequence
}
\examples{
protein_df <- translateDNA(DNA_df)

}
